+++ tests/browser_modules/requests_filter.test.js
@@ -170,4 +170,23 @@ describe("Requests Filter", function() {
         const afterFilter2 = await browser.requests.filter.url(/api/).responseBody("not response").requests;
         assert.strictEqual(afterFilter2.length, 0);
     });
+
+    it("Filter By Pending", async() => {
+        const response = Object.assign({
+            auto: false,
+            body: {result: "DUMMY"}
+        });
+        const mock = browser.requests.mock(configUrls.api, response);
+        const pending1 = await browser.requests.filter.pending().requests;
+        assert.strictEqual(pending1.length, 0);
+
+        await browser.clickText("click me");
+        await browser.wait(10);
+        const pending2 = await browser.requests.filter.pending().requests;
+        assert.strictEqual(pending2.length, 1);
+        mock.trigger();
+        await browser.wait(10);
+        const pending3 = await browser.requests.filter.pending().requests;
+        assert.strictEqual(pending3.length, 0);
+    });
 });